debt - allow to track disposables in running app
